perf(export): lazy-load docx/pptxgenjs/write-excel-file out of the eager bundle
exportFormats.js statically imported docx, pptxgenjs, and write-excel-file,
and was pulled into the initial bundle via ExportDialog and the
buildChatExportFilename/buildChatExportTitle helpers re-exported through
api/endpoints/apps.js — so every user downloaded ~800KB of export libraries
they'd only need if they opened the export dialog.

Split the dependency-free filename/title helpers into exportNaming.js so
apps.js can import them without pulling in exportFormats.js, switched the
docx/pptxgenjs/write-excel-file usages inside exportFormats.js to dynamic
imports, and made ExportDialog dynamically import exportFormats.js per
format instead of statically. Fixes #1770.

Note: the issue's claim that the 'xlsx' dependency was dead is incorrect —
it's used via a dynamic import('xlsx') in
features/upload/utils/fileProcessing.js for reading uploaded spreadsheets
(distinct from write-excel-file, used for exports), so it was left in place.
