+++ client/src/utils/exportFormats.js
@@ -1,135 +1,25 @@
-import writeXlsxFile from 'write-excel-file';
-import { Document, Paragraph, TextRun, HeadingLevel, Table, TableRow, TableCell } from 'docx';
-import PptxGenJS from 'pptxgenjs';
+// docx/pptxgenjs/write-excel-file are heavy and only needed when a user
+// actually exports to one of those formats — they're dynamically imported
+// inside the exportTo* functions below instead of statically here, so this
+// module (and everything that imports it) doesn't pull them into the eager
+// bundle. Filename/title helpers live in exportNaming.js, which has no heavy
+// deps, so callers that only need those (e.g. api/endpoints/apps.js) can
+// import from there directly.
+export {
+  sanitizeForSpreadsheet,
+  slugifyForFilename,
+  getChatTopicSlug,
+  formatDateTimeForFilename,
+  formatDateTimeForTitle,
+  buildChatExportFilename,
+  buildChatExportTitle
+} from './exportNaming';
 
-/**
- * Filename- and title-building helpers shared by all chat export formats.
- *
- * Goal: replace generic "chat-export-2026-06-09T15-30-00.xlsx" filenames
- * and "Chat Export - iHub Apps" document titles with something meaningful
- * — the app name, a topic slug derived from the first user message, and a
- * readable date.
- */
-
-/**
- * Neutralize spreadsheet formula injection (OWASP CSV injection guidance).
- * Values starting with =, +, -, @, tab, or CR are interpreted as formulas by
- * Excel/LibreOffice; prefixing with a single quote forces them to render as
- * plain text instead of executing.
- */
-export const sanitizeForSpreadsheet = value => {
-  if (value === null || value === undefined) return '';
-  const stringValue = String(value);
-  return /^[=+\-@\t\r]/.test(stringValue) ? `'${stringValue}` : stringValue;
-};
-
-/** Strip markdown noise, collapse whitespace, ASCII-kebab-case, cap length. */
-export const slugifyForFilename = (text, maxChars = 40) => {
-  if (!text || typeof text !== 'string') return '';
-  return text
-    .replace(/```[\s\S]*?```/g, ' ') // drop fenced code
-    .replace(/`[^`]*`/g, ' ') // drop inline code
-    .replace(/!\[[^\]]*\]\([^)]*\)/g, ' ') // drop images
-    .replace(/\[([^\]]+)\]\([^)]*\)/g, '$1') // unwrap links
-    .replace(/[*_~#>]/g, ' ') // drop markdown markers
-    .normalize('NFKD')
-    .replace(/[̀-ͯ]/g, '') // strip accents
-    .replace(/[^a-zA-Z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '')
-    .toLowerCase()
-    .slice(0, maxChars)
-    .replace(/-+$/, '');
-};
-
-/**
- * Derive a short topic slug from the first non-greeting user message.
- * Returns '' when the chat has no user content to summarize from.
- */
-export const getChatTopicSlug = messages => {
-  if (!Array.isArray(messages)) return '';
-  const firstUser = messages.find(m => m && m.role === 'user' && !m.isGreeting && m.content);
-  if (!firstUser) return '';
-  return slugifyForFilename(firstUser.content, 40);
-};
-
-const pad2 = n => String(n).padStart(2, '0');
-
-/** `2026-06-09_1530` — filesystem-safe, sortable. */
-export const formatDateTimeForFilename = (date = new Date()) =>
-  `${date.getFullYear()}-${pad2(date.getMonth() + 1)}-${pad2(date.getDate())}_` +
-  `${pad2(date.getHours())}${pad2(date.getMinutes())}`;
-
-/** `2026-06-09 15:30` — human-readable, used in document titles. */
-export const formatDateTimeForTitle = (date = new Date()) =>
-  `${date.getFullYear()}-${pad2(date.getMonth() + 1)}-${pad2(date.getDate())} ` +
-  `${pad2(date.getHours())}:${pad2(date.getMinutes())}`;
-
-/**
- * Build a descriptive download filename.
- *
- *   Full chat with topic:  `sales-assistant-pricing-discussion-2026-06-09_1530.docx`
- *   Full chat, no topic:   `sales-assistant-chat-2026-06-09_1530.docx`
- *   Single message:        `sales-assistant-message-2026-06-09_1530.docx`
- *   No app context:        `chat-2026-06-09_1530.docx`
- */
-export const buildChatExportFilename = ({
-  format,
-  appName,
-  appId,
-  messages,
-  isSingleMessage = false,
-  date = new Date()
-}) => {
-  const ext = (format || '').toLowerCase();
-  const appSlug = slugifyForFilename(appId || appName || '', 30);
-  const dateStr = formatDateTimeForFilename(date);
-
-  let middle;
-  if (isSingleMessage) {
-    middle = 'message';
-  } else {
-    const topic = getChatTopicSlug(messages);
-    middle = topic || 'chat';
-  }
-
-  const parts = [appSlug, middle, dateStr].filter(Boolean);
-  return `${parts.join('-')}.${ext}`;
-};
-
-/**
- * Build a descriptive in-document title.
- *
- *   `Sales Assistant — Pricing Discussion (2026-06-09 15:30)`
- *   `Sales Assistant — Message (2026-06-09 15:30)`
- *   `Sales Assistant — Chat (2026-06-09 15:30)`
- */
-export const buildChatExportTitle = ({
-  appName,
-  messages,
-  isSingleMessage = false,
-  date = new Date()
-}) => {
-  const dateStr = formatDateTimeForTitle(date);
-  const app = appName || 'iHub Apps';
-
-  if (isSingleMessage) return `${app} — Message (${dateStr})`;
-
-  // Use the first user message as a short topic — capitalize words, cap length.
-  const firstUser = Array.isArray(messages)
-    ? messages.find(m => m && m.role === 'user' && !m.isGreeting && m.content)
-    : null;
-  if (firstUser?.content) {
-    const topic = firstUser.content
-      .replace(/```[\s\S]*?```/g, ' ')
-      .replace(/`[^`]*`/g, ' ')
-      .replace(/[*_~#>]/g, ' ')
-      .replace(/\s+/g, ' ')
-      .trim()
-      .slice(0, 60);
-    if (topic) return `${app} — ${topic} (${dateStr})`;
-  }
-  return `${app} — Chat (${dateStr})`;
-};
+import {
+  buildChatExportFilename,
+  buildChatExportTitle,
+  sanitizeForSpreadsheet
+} from './exportNaming';
 
 /**
  * Parse inline markdown formatting (bold, italic, code) within a text line
@@ -386,7 +276,8 @@ const parseMarkdown = content => {
 /**
  * Convert parsed markdown blocks to DOCX paragraphs
  */
-const markdownToDOCX = blocks => {
+const markdownToDOCX = (docxLib, blocks) => {
+  const { Paragraph, TextRun, HeadingLevel, Table, TableRow, TableCell } = docxLib;
   const paragraphs = [];
 
   blocks.forEach(block => {
@@ -700,6 +591,7 @@ export const exportToXLSX = async (
   const columns = [{ width: 15 }, { width: 20 }, { width: 80 }];
 
   // Write XLSX file
+  const { default: writeXlsxFile } = await import('write-excel-file');
   await writeXlsxFile(data, {
     columns,
     fileName: filename
@@ -794,6 +686,10 @@ export const exportToDOCX = async (
   });
   const docTitle = buildChatExportTitle({ appName, messages, isSingleMessage });
 
+  const docxLib = await import('docx');
+  const { Document, Paragraph, TextRun, HeadingLevel, Packer, AlignmentType, convertInchesToTwip } =
+    docxLib;
+
   const children = [];
 
   // Add title
@@ -861,7 +757,7 @@ export const exportToDOCX = async (
 
     // Parse markdown and convert to DOCX paragraphs
     const blocks = parseMarkdown(content);
-    const parsedParagraphs = markdownToDOCX(blocks);
+    const parsedParagraphs = markdownToDOCX(docxLib, blocks);
     children.push(...parsedParagraphs);
 
     children.push(new Paragraph({ text: '' }));
@@ -916,7 +812,6 @@ export const exportToDOCX = async (
   }
 
   // Create document with proper numbering support
-  const { AlignmentType, convertInchesToTwip } = await import('docx');
   const doc = new Document({
     numbering: {
       config: [
@@ -946,7 +841,6 @@ export const exportToDOCX = async (
   });
 
   // Use docx Packer to generate blob
-  const { Packer } = await import('docx');
   const blob = await Packer.toBlob(doc);
 
   // Download file
@@ -1044,6 +938,7 @@ export const exportToPPTX = async (
   });
   const docTitle = buildChatExportTitle({ appName, messages, isSingleMessage });
 
+  const { default: PptxGenJS } = await import('pptxgenjs');
   const pres = new PptxGenJS();
 
   // Title slide
